Move common serializer behaviour in base class
