+++ lib/dynameister/serializers.rb
@@ -1,3 +1,4 @@
+require_relative "serializers/base_item_serializer"
 require_relative "serializers/get_item_serializer"
 require_relative "serializers/delete_item_serializer"
 require_relative "serializers/put_item_serializer"
