+++ lib/dynameister/serializers/base_item_serializer.rb
@@ -1,0 +1,25 @@
+module Dynameister
+
+  module Serializers
+
+    class BaseItemSerializer
+
+      attr_accessor :key
+
+      def initialize(table_name:, key:)
+        @table_name = table_name
+        @key = key
+      end
+
+      def to_h
+        {
+          table_name: @table_name,
+          key:        key
+        }
+      end
+
+    end
+
+  end
+
+end
