+++ lib/dynameister/serializers/delete_item_serializer.rb
@@ -2,23 +2,7 @@ module Dynameister
 
   module Serializers
 
-    class DeleteItemSerializer
-
-      attr_accessor :key
-
-      def initialize(table_name:, key:)
-        @table_name = table_name
-        @key = key
-      end
-
-      def to_h
-        {
-          table_name: @table_name,
-          key:        key
-        }
-      end
-
-    end
+    class DeleteItemSerializer < BaseItemSerializer; end
 
   end
